requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
django-nested-admin==4.0.2
idna==3.4
python-monkey-business==1.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
psycopg2-binary==2.9.10
django-dump-load-utf8==0.0.4
python-dotenv==1.0.1

elasticsearch == 8.17.2
elasticsearch-dsl == 8.17.1

djangorestframework == 3.15.2